requirements.txt
absl-py==0.12.0
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.1
google-auth-oauthlib==0.4.4
grpcio==1.37.0
gym==0.21.0
idna==2.10
iniconfig==1.1.1
jsonpickle==1.5.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
munch==2.5.0
networkx==2.5.1
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.15.8
py==1.10.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacred==0.8.2
scipy==1.6.2
six==1.15.0
smmap==4.0.0
stable-baselines3==1.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
toml==0.10.2
torch==1.8.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1

rware==1.0.3
lbforaging==1.1.1
imageio-ffmpeg==0.5.1
